Painting Infrastructure
Created a "ColorManager" script and added it to a "Gameplay Manager" object. This handles input relating to color selection, application and removal. It holds a publicly accessible list of all colors used in gameplay and the objects that are currently painted.

Created a "Colorable" script that handles setting the object's color in accordance to the "ColorManager". It dynamically adds or removes the correct scripts to allow for the functionality of each color.

Added 3 scripts named "Red", "Green" and "Blue". They are the base on which each color's functionality will be built. They have to be named the same as the material which is applied to the objects.