Green functionality, bug fixes & organizational changes.
Green objects freeze in place but retain their velocity. When unfrozen, they'll continue their trajectory.

Added a bool overrideColors to ColorManager. If disabled, the player is forced to remove an object's color before they can paint it a different color.

The Interactable script was casting a ray from the player's feet. It now casts from the center of the main camera.

Applied changes from past commit to Player prefab.
The keybind for each GameplayColor is now accessible in the inspector.
ColorManager now has a default material for unpainted objects to return to.