Brush, UI and URP
Modeled, animated and textured a brush model.

Added Jiggle physics to the brush using a plugin (EZSoftBone by EZHex1991).

Added a crosshair at the center of the screen for easier aiming.

Added a script called BrushController.cs that makes sure the brush physics and texture are correct, and also updates the color of the crosshair.

Hooked up the brush animations to an animator controller and updated ColorManager.cs to interface with it.

Added a secondary camera to render the brush on its own layer thus avoiding clipping when the player gets too close to a wall.

Upgraded the Render Pipeline to Unity's Universal Render Pipeline (URP). This allows for post processing effects and layering the default and brush views together.

Added a pause menu accessible through the Escape key and a UIManager.cs to handle all things UI.

Imported TextMeshPro to add text to the UI.