Blue and color unlocking
Added moving platforms to the game. It took 27 hours of work spanning 3 days. I was sick while doing it.

Blue.cs is the brain of the platform. It does most of the collision checks to change direction.
PlatformRideTrigger.cs lets objects ride the platform by adding them as children to the platform's transform.
Obstacle.cs builds a chain of obstacles in front of the platform that propagates collisions to avoid clipping objects into walls.

There are issues with it. I have noted them in the comments.

Exiting play mode while a blue platform is active throws errors. I do not care. They don't affect the game.

ColorManager.cs now has logic to deny switching to a color that hasn't been unlocked yet.

The demo scene is a mess because I am too.