Fixed small bugs & tutorial showcase extension
Made Red have an arbitrary offset from the player when held instead of always sticking to the center of the screen.

Fixed button so it can work even when upside down (an interaction with RideTrigger was breaking it, therefore all button elements must now be tagged "Button")

Extended the Tutorial scene to include the rest of the tutorial puzzles.

Wrote SignalAccumulator.cs to manage the opening and closing of doors that need more than one button to be pressed before opening.

Wrote ColorUnlock.cs that can be added to a trigger object so that a color will be unlocked.

Made all colorable cubes respawn when dropped bellow -10 units (off the map)