Finished wall splatting code. Need feedback.
Finished the wall splatting code.

This approach requires very careful consideration of each new object we add to the scene.

All geometry must have non-overlapping UVs, or at least a second UV channel where no face is overlapping.

I won't be surprised if we go a different direction because of that.

I've made a prefab of a cube with correct UVs, though stretching would still be an issue. It seems that for this to work, all levels must be designed in blender.

again, you tell me if this is okay.

The reason to not use decals is that the player won't be able to draw infinite decals, whereas wit this approach, all painting is computationally equivalent.