Wall splatting proof of concept
Added a "Splat" layer for all our splat catchers.

SplatPainter.cs will generate a texture for each face in the Splat layer once the player tries to paint it. After that, it will update the texture accordingly.

A copy of each visible face is needed for the effect to work. It is possible to do this relatively easily AFTER the levels are made, using the "FBX Exporter" package ,and Blender to inset all the faces evenly.