Added Jukebox
The jukebox can loop through a playlist of tracks, shuffle, repeat and pause.

It can also fade from one track to another, or transition to a completely new track before continuing with the playlist.

The jukebox is a singleton GameObject, meaning that it will override any new jukebox it finds when moving to a new scene. Therefore, if you want to reference the jukebox, use "FindFirstObjectByType<MusicController>();"

Tracks should be preloaded for seamless playback. Three demo tracks have been added to the project. They may be removed.