Organization is NOT optional when collaborating.
why was the canvas prefab split in pieces? why did the whiteboard have a canvas renderer? why were the sliders not connected to their scripts? why was nothing named appropriately? you know I'll have to parse this nonsense if I am to add to it.

brother, it's been 2 hours. I was supposed to finish the green level.

please organize your things before pushing.